add crud for users, add AllowService to check if user is allowed to make action

+++ src/Controller/CourseController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Course;
 use App\Form\CourseType;
 use App\Repository\CourseRepository;
+use App\Service\AllowService;
 use App\Service\CourseService;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -68,8 +69,14 @@ class CourseController extends AbstractController
      * @Route("/{id}/edit", methods={"GET","POST"})
      * @IsGranted("ROLE_CLIENT")
      */
-    public function edit(Request $request, Course $course): Response
+    public function edit(Request $request, Course $course, AllowService $service): Response
     {
+        if (!$service->allowAccess($this->getUser(), $course->getCreator())) {
+            $this->addFlash('danger', 'Vous n\' avez pas accès à ces fonctions');
+
+            return $this->redirectToRoute('app_homepage_index');
+        }
+
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
 
@@ -92,8 +99,14 @@ class CourseController extends AbstractController
      * @Route("/{id}", methods={"DELETE"})
      * @IsGranted("ROLE_CLIENT")
      */
-    public function delete(Request $request, Course $course): Response
+    public function delete(Request $request, Course $course, AllowService $service): Response
     {
+        if (!$service->allowAccess($this->getUser(), $course->getCreator())) {
+            $this->addFlash('danger', 'Vous n\' avez pas accès à ces fonctions');
+
+            return $this->redirectToRoute('app_homepage_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($course);
